+++ src/data/import_phenotypes.py
@@ -13,7 +13,7 @@ from click_option_group import optgroup, RequiredMutuallyExclusiveOptionGroup
 from pathlib import Path
 import pandas as pd
 
-from src.features.smarterdb import global_connection, Dataset, Phenotype
+from src.features.smarterdb import global_connection, Dataset, Phenotype, Breed
 from src.data.common import (
     fetch_and_check_dataset, pandas_open, get_sample_species)
 from src.features.utils import sanitize
@@ -101,6 +101,12 @@ def add_phenotype_by_breed(
     for index, row in data.iterrows():
         logger.debug(f"{index}, {row}")
         breed = row.get(columns["breed_column"])
+
+        # check that this breed exists
+        qs = Breed.objects.filter(name=breed, species=dst_dataset.species)
+
+        if qs.count() != 1:
+            raise Exception(f"Breed '{breed}' not found in database!")
 
         # get columns modelled in smarter database
         named_columns = get_named_columns(row, columns, breed)